# pkg/swagger-ui/swagger.json
{
  "swagger": "2.0",
  "info": {},
  "paths": {
...
    "/host/{id}": {
      "get": {
        "description": "Fetch a Host",
        "tags": [
          "host-tag"
        ],
        "operationId": "idOfFetchHost",
        "responses": {
          "200": {
            "$ref": "#/responses/fetchHostResponse"
          },
          "500": {
            "$ref": "#/responses/DefaultResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/LambdatestIncPrivate/bookkeeping/pkg/service"
    },
...
    "Host": {
      "type": "object",
      "properties": {
        "Capacity": {
          "type": "integer",
          "format": "int64"
        },
        "ClockSpeed": {
          "type": "integer",
          "format": "int64"
        },
        "Hash": {
          "type": "string"
        },
        "available_capacity": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AvailableCapacity"
        },
        "available_storage": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AvailableStorage"
        },
        "available_templates": {
          "type": "string",
          "x-go-name": "AvailableTemplates"
        },
        "cores_limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CPULimit"
        },
        "cpu": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Cores"
        },
        "created_by": {
          "type": "string",
          "x-go-name": "CreatedBy"
        },
        "created_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created"
        },
        "fqdn": {
          "type": "string",
          "x-go-name": "Fqdn"
        },
        "health_status": {
          "type": "string",
          "x-go-name": "HealthStatus"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "org_id": {
          "type": "string",
          "x-go-name": "OrganizationID"
        },
        "os_name": {
          "type": "string",
          "x-go-name": "OSName"
        },
        "private_ip": {
          "type": "string",
          "x-go-name": "PrivateIP"
        },
        "provider": {
          "type": "string",
          "x-go-name": "Provider"
        },
        "public_ip": {
          "type": "string",
          "x-go-name": "PublicIP"
        },
        "ram": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RAM"
        },
        "ram_limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RAMLimit"
        },
        "region": {
          "type": "string",
          "x-go-name": "Region"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "storage": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Storage"
        },
        "storage_limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StorageLimit"
        },
        "updated_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Updated"
        }
      },
      "x-go-package": "github.com/LambdatestIncPrivate/bookkeeping/pkg/models"
    },
...
  },
  "responses": {
...
    "DefaultResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/DefaultResponse"
      }
    },
...
    "fetchHostResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Host"
      }
    },
...
  }
}